working integration tests using both seed and JSON input
