--- integration_tests/test_example_failing.py
@@ -1,70 +1,0 @@
-import pytest
-from dbt.tests.util import run_dbt
-
-# our file contents
-from fixtures import (
-    my_seed_csv,
-    my_model_sql,
-    my_model_yml,
-)
-
-# class must begin with 'Test'
-class TestExample:
-    """
-    Methods in this class will be of two types:
-    1. Fixtures defining the dbt "project" for this test case.
-       These are scoped to the class, and reused for all tests in the class.
-    2. Actual tests, whose names begin with 'test_'.
-       These define sequences of dbt commands and 'assert' statements.
-    """
-    
-    # configuration in dbt_project.yml
-    @pytest.fixture(scope="class")
-    def project_config_update(self):
-        return {
-          "name": "example",
-          "models": {"+materialized": "view"}
-        }
-
-    # everything that goes in the "seeds" directory
-    @pytest.fixture(scope="class")
-    def seeds(self):
-        return {
-            "my_seed.csv": my_seed_csv,
-        }
-
-    # everything that goes in the "models" directory
-    @pytest.fixture(scope="class")
-    def models(self):
-        return {
-            "my_model.sql": my_model_sql,
-            "my_model.yml": my_model_yml,
-        }
-        
-        # continued from above
-
-    # The actual sequence of dbt commands and assertions
-    # pytest will take care of all "setup" + "teardown"
-    def test_run_seed_test(self, project):
-        """
-        Seed, then run, then test. We expect one of the tests to fail
-        An alternative pattern is to use pytest "xfail" (see below)
-        """
-        # seed seeds
-        results = run_dbt(["seed"])
-        assert len(results) == 1
-        # run models
-        results = run_dbt(["run"])
-        assert len(results) == 1
-        # test tests
-        results = run_dbt(["test"], expect_pass = False) # expect failing test
-        assert len(results) == 2
-        # validate that the results include one pass and one failure
-        result_statuses = sorted(r.status for r in results)
-        assert result_statuses == ["fail", "pass"]
-
-    @pytest.mark.xfail
-    def test_build(self, project):
-        """Expect a failing test"""
-        # do it all
-        results = run_dbt(["build"])
