+++ integration_tests/test_stg_ga4__event_to_query_string_params.py
@@ -1,0 +1,47 @@
+import pytest
+from dbt.tests.util import read_file,check_relations_equal,run_dbt
+from base_unit_test import BaseUnitTestModel
+
+SOURCE_JSON = """
+{  "event_key": "aD8bvMcixjJbYvx61zWbEw\u003d\u003d", "page_query_string": "param1=value1&param2=value2"}
+""".lstrip()
+
+EXPECTED_CSV = """event_key,param,value
+aD8bvMcixjJbYvx61zWbEw\u003d\u003d,param1,value1
+aD8bvMcixjJbYvx61zWbEw\u003d\u003d,param2,value2
+""".lstrip()
+
+actual = read_file('../models/staging/ga4/stg_ga4__event_to_query_string_params.sql').replace(
+    "ref('stg_ga4__events')",
+    "source('fixture','SOURCE_JSON')"
+)
+
+models__config_yml = """
+version: 2
+sources:
+  - name: fixture
+    schema: "{{ target.schema }}"
+    tables:
+      - name: SOURCE_JSON
+"""
+
+class TestUnitTestComplexModel(BaseUnitTestModel):
+    # everything that goes in the "seeds" directory (= CSV format)
+    @pytest.fixture(scope="class")
+    def seeds(self):
+        return {
+            "expected.csv": EXPECTED_CSV,
+        }
+
+    # everything that goes in the "models" directory (= SQL)
+    @pytest.fixture(scope="class")
+    def models(self):
+        return {
+            "actual.sql": actual,
+            "config.yml": models__config_yml
+        }
+    
+    def test_mock_run_and_check(self, project):
+        self.upload_json_fixture(project, "source.json", SOURCE_JSON, "SOURCE_JSON" )
+        run_dbt(["build"])
+        check_relations_equal(project.adapter, ["actual", "expected"])
